rgbd_tracker/CLGSD/requirements_conda.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
absl-py=0.9.0=pypi_0
addict=2.2.1=pypi_0
astor=0.8.1=pypi_0
attrs=19.3.0=pypi_0
blas=1.0=mkl
ca-certificates=2020.1.1=0
cachetools=4.0.0=pypi_0
certifi=2020.4.5.1=py37_0
cffi=1.14.0=pypi_0
chardet=3.0.4=pypi_0
click=7.1.1=pypi_0
cloudpickle=1.3.0=pypi_0
colorama=0.4.3=pypi_0
cudatoolkit=10.0.130=0
cudnn=7.6.5=cuda10.0_0
cycler=0.10.0=py37_0
cython=0.29.15=py37he6710b0_0
dash=1.10.0=pypi_0
dash-core-components=1.9.0=pypi_0
dash-html-components=1.0.3=pypi_0
dash-renderer=1.3.0=pypi_0
dash-table=4.6.2=pypi_0
dbus=1.13.12=h746ee38_0
dcnv2=0.1=dev_0
decorator=4.4.2=pypi_0
detectron2=0.1.1=pypi_0
easydict=1.9=pypi_0
expat=2.2.6=he6710b0_0
flask=1.1.2=pypi_0
flask-compress=1.4.0=pypi_0
fontconfig=2.13.0=h9420a91_0
freetype=2.9.1=h8a8886c_1
future=0.18.2=pypi_0
fvcore=0.1.dev200420=pypi_0
gast=0.2.2=pypi_0
glib=2.63.1=h5a9c865_0
google-auth=1.11.3=pypi_0
google-auth-oauthlib=0.4.1=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.27.2=pypi_0
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb453b48_1
h5py=2.10.0=pypi_0
icu=58.2=h9c2bf20_1
idna=2.9=pypi_0
imagecorruptions=1.1.0=pypi_0
imageio=2.8.0=pypi_0
imgaug=0.2.6=pypi_0
importlib-metadata=1.6.0=pypi_0
intel-openmp=2020.0=166
itsdangerous=1.1.0=pypi_0
jinja2=2.11.1=pypi_0
joblib=0.14.1=pypi_0
jpeg=9b=h024ee3a_2
jpeg4py=0.1.4=pypi_0
jsonpatch=1.25=pypi_0
jsonpointer=2.0=pypi_0
jsonschema=3.2.0=pypi_0
keras=2.3.1=pypi_0
keras-applications=1.0.8=pypi_0
keras-preprocessing=1.1.0=pypi_0
kiwisolver=1.1.0=py37he6710b0_0
ld_impl_linux-64=2.33.1=h53a641e_7
libedit=3.1.20181209=hc058e9b_0
libffi=3.2.1=hd88cf55_4
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_0
libuuid=1.0.3=h1bed415_2
libxcb=1.13=h1bed415_1
libxml2=2.9.9=hea5a465_1
llvmlite=0.31.0=pypi_0
markdown=3.2.1=pypi_0
markupsafe=1.1.1=pypi_0
matplotlib=3.1.3=py37_0
matplotlib-base=3.1.3=py37hef1b27d_0
mkl=2020.0=166
mkl-service=2.3.0=py37he904b0f_0
mkl_fft=1.0.15=py37ha843d7b_0
mkl_random=1.1.0=py37hd6b4f25_0
mmcv=0.2.16=pypi_0
mmdet=1.0.0+56db9d2=dev_0
mock=4.0.2=pypi_0
ncurses=6.2=he6710b0_0
networkx=2.4=pypi_0
ninja=1.9.0=py37hfd86e86_0
nms=0.1.6=pypi_0
numba=0.48.0=pypi_0
numpy=1.18.1=py37h4f9e942_0
numpy-base=1.18.1=py37hde5b4d6_1
oauthlib=3.1.0=pypi_0
olefile=0.46=py37_0
opencv-python=4.2.0.32=pypi_0
openssl=1.1.1g=h7b6447c_0
opt-einsum=3.2.1=pypi_0
ordered-set=3.1.1=pypi_0
pandas=1.0.1=py37h0573a6f_0
pcre=8.43=he6710b0_0
pillow=6.2.2=pypi_0
pip=20.0.2=py37_1
plotly=4.6.0=pypi_0
portalocker=1.7.0=pypi_0
progress=1.5=pypi_0
progressbar=2.5=pypi_0
protobuf=3.11.3=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycocotools=2.0.0=pypi_0
pycparser=2.20=pypi_0
pydot=1.4.1=pypi_0
pylatex=1.3.1=pypi_0
pyparsing=2.4.6=py_0
pyqt=5.9.2=py37h05f1152_2
pyrsistent=0.16.0=pypi_0
python=3.7.6=h0371630_2
python-dateutil=2.8.1=py_0
pytorch=1.4.0=py3.7_cuda10.0.130_cudnn7.6.3_0
pytz=2019.3=py_0
pywavelets=1.1.1=pypi_0
pyzmq=19.0.0=pypi_0
qt=5.9.7=h5867ecd_1
readline=7.0=h7b6447c_5
requests=2.23.0=pypi_0
requests-oauthlib=1.3.0=pypi_0
retrying=1.3.3=pypi_0
roi-align=0.0.1=pypi_0
rsa=4.0=pypi_0
scikit-image=0.16.2=pypi_0
scikit-learn=0.22.2.post1=pypi_0
scipy=1.4.1=pypi_0
seaborn=0.10.0=pypi_0
setuptools=46.0.0=py37_0
shapely=1.7.0=pypi_0
sip=4.19.8=py37hf484d3e_0
six=1.14.0=py37_0
sqlite=3.31.1=h7b6447c_0
tabulate=0.8.7=pypi_0
tb-nightly=2.2.0a20200315=pypi_0
tensorboard=1.15.0=pypi_0
tensorboard-plugin-wit=1.6.0.post2=pypi_0
tensorboardx=2.0=pypi_0
tensorflow-estimator=1.15.1=pypi_0
tensorflow-gpu=1.15.0=pypi_0
termcolor=1.1.0=pypi_0
tk=8.6.8=hbc83047_0
torchfile=0.1.0=pypi_0
torchvision=0.5.0=py37_cu100
tornado=6.0.4=py37h7b6447c_1
tqdm=4.32.0=pypi_0
urllib3=1.25.8=pypi_0
visdom=0.1.8.9=pypi_0
vot-toolkit=0.2.0=pypi_0
vot-trax=3.0.1=pypi_0
websocket-client=0.57.0=pypi_0
werkzeug=1.0.0=pypi_0
wheel=0.34.2=py37_0
wrapt=1.12.1=pypi_0
xz=5.2.4=h14c3975_4
yacs=0.1.6=pypi_0
zipp=3.1.0=pypi_0
zlib=1.2.11=h7b6447c_3
zstd=1.3.7=h0b5b093_0
